+++ inc/lilyParser/index.ts
@@ -2,7 +2,7 @@
 import jison from "jison";
 
 import LilyDocument from "./lilyDocument";
-import {termDictionary, getDurationSubdivider} from "./lilyDocument";
+import {termDictionary, getDurationSubdivider} from "./lilyTerms";
 
 import * as measures from "./measures";
 import {createPianoRhythm} from "./pianoRhythm";