+++ inc/lilyParser/lilyInterpreter.ts
@@ -1,10 +1,12 @@
 
+import {romanize} from "../romanNumeral";
 import {WHOLE_DURATION_MAGNITUDE} from "./utils";
-import {parseRaw, Primitive} from "./lilyTerms";
+import {parseRaw} from "./lilyTerms";
 
 // eslint-disable-next-line
-import {BaseTerm, Root, Block, MusicEvent, Repeat, Relative, TimeSignature, Partial, Times, Tuplet, Grace, Clef, KeySignature, OctaveShift, Duration, ChordElement, Chord, MusicBlock, Assignment, Variable, Command, SimultaneousList, ContextedMusic} from "./lilyTerms";
-
+import {BaseTerm, Root, Block, MusicEvent, Repeat, Relative, TimeSignature, Partial, Times, Tuplet, Grace, Clef, KeySignature, OctaveShift, Duration, ChordElement, Chord, MusicBlock, Assignment, Variable, Command, SimultaneousList, ContextedMusic, Primitive, Version, Scheme, Include} from "./lilyTerms";
+// eslint-disable-next-line
+import LilyDocument from "./lilyDocument";
 
 
 interface DurationContextStackStatus {
@@ -196,18 +198,41 @@ export default class LilyInterpreter {
 
 	musicTracks: MusicBlock[] = [];
 
+	version: Version = null;
+	header: Block = null;
+	statements: BaseTerm[] = [];
+	paper: Block = null;
+	layout: Block = null;
+	score: Block = null;
 
-	eval (term: BaseTerm): BaseTerm {
+
+	static trackName (index: number): string {
+		return `Voice_${romanize(index)}`;
+	};
+
+
+	/*eval (term: BaseTerm): BaseTerm {
 		return this.execute(term.clone());
-	}
+	}*/
 
 
-	interpretMusic (music: BaseTerm) {
+	interpretMusic (music: BaseTerm): Variable {
 		//console.log("interpretMusic:", music);
 		const context = new StaffContext();
 		context.execute(music);
 
 		this.musicTracks.push(context.track);
+
+		const varName = LilyInterpreter.trackName(this.musicTracks.length);
+
+		return new Variable({name: varName});
+	}
+
+
+	interpretDocument (doc: LilyDocument): this {
+		this.execute(doc.root);
+
+		return this;
 	}
 
 
@@ -216,17 +241,47 @@ export default class LilyInterpreter {
 			return term;
 
 		if (term instanceof Root) {
-			for (const section of term.sections) 
-				this.execute(section, {execMusic: true});
+			for (const section of term.sections) {
+				const sec = this.execute(section, {execMusic: true});
+				if (sec instanceof Version)
+					this.version = sec;
+				else if (sec instanceof Scheme)
+					this.statements.push(sec);
+				else if (sec instanceof Include)
+					this.statements.push(sec);
+				else if (sec instanceof Block) {
+					switch (sec.head) {
+					case "\\header":
+						this.header = sec;
+
+						break;
+					case "\\paper":
+						this.paper = sec;
+
+						break;
+					case "\\layout":
+						this.layout = sec;
+
+						break;
+					case "\\score":
+						this.score = sec;
+
+						break;
+					}
+				}
+			}
 		}
 		else if (term instanceof Assignment) {
 			const value = this.execute(term.value);
 			this.variableTable.set(term.key as string, value);
 		}
 		else if (term instanceof Block) {
-			if (term.block === "score") {
-				for (const subterm of term.body) 
-					this.execute(subterm, {execMusic: true});
+			if (term.head === "\\score") {
+				return new Block({
+					block: term.block,
+					head: term.head,
+					body: term.body.map(subterm => this.execute(subterm, {execMusic: true})),
+				});
 			}
 		}
 		else if (term instanceof Variable) {
@@ -238,8 +293,10 @@ export default class LilyInterpreter {
 		}
 		else if (term instanceof MusicBlock) {
 			const result = new MusicBlock({body: term.body.map(subterm => this.execute(subterm)).filter(term => term)});
-			if (execMusic)
-				this.interpretMusic(result);
+			if (execMusic) {
+				const variable = this.interpretMusic(result);
+				return new MusicBlock({body: [variable]});
+			}
 
 			return result;
 		}
@@ -247,14 +304,24 @@ export default class LilyInterpreter {
 			return new SimultaneousList({list: term.list.map(subterm => this.execute(subterm, {execMusic})).filter(term => term)});
 		else if (term instanceof ContextedMusic)
 			return new ContextedMusic({head: this.execute(term.head), lyrics: this.execute(term.lyrics), body: this.execute(term.body, {execMusic})});
-		else if (term instanceof Command) {
-			const result = parseRaw({proto: term.proto, cmd: term.cmd, args: term.args.map(arg => this.execute(arg))});
-			if (execMusic && result.isMusic)
-				this.interpretMusic(result);
-
-			return result;
-		}
+		else if (term instanceof Command)
+			return parseRaw({proto: term.proto, cmd: term.cmd, args: term.args.map(arg => this.execute(arg, {execMusic}))});
 
 		return term;
+	}
+
+
+	toDocument (): LilyDocument {
+		const root = new Root({sections: [
+			this.version,
+			this.header,
+			...this.statements,
+			// TODO: insert track assignments
+			this.paper,
+			this.layout,
+			this.score,
+		].filter(section => section)});
+
+		return new LilyDocument(root);
 	}
 };